requirements.txt
aiohttp==3.7.4.post0
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.5.7
bidict==0.21.2
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
fastapi==0.66.0
h11==0.12.0
httptools==0.2.0
idna==2.10
multidict==5.1.0
paho-mqtt==1.5.1
psutil==5.8.0
pycodestyle==2.7.0
pydantic==1.8.2
PySide2==5.15.2
python-dotenv==0.18.0
python-engineio==4.2.0
python-socketio==5.3.0
PyYAML==5.4.1
requests==2.25.1
shiboken2==5.15.2
starlette==0.14.2
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.5
uvicorn==0.14.0
uvloop==0.15.2
watchgod==0.7
websocket-client==1.1.0
websockets==9.1
yarl==1.6.3
